+++ src/gen-msg.cc
@@ -85,10 +85,10 @@ gen_msg (int argc, char **argv)
 
       printf ("static const char *const message_string[] =\n");
       printf ("{\n");
-      int l = 0;
+      __int64 l = 0;
       for (int i = 0; i < numberof (msg); i++)
         {
-          printf ("  SSM + %d,\n", l);
+          printf ("  SSM + %I64d,\n", l);
           l += strlen (msg[i].text) + 1;
         }
       printf ("};\n\n");
